Modified the way colors where added to outputStyledDocument in DragonConsole and made it much easier for new Styles to be added "on the fly." All styles are now controlled by the DragonConsole and maintained in an ArrayList<TextColor> textColors so that new styles can easily have each background and foreground color previously added.
I've moved the DocumentStyler from the default dragonconsole package into dragonconsole.util where it made more sense to be as it is a utility class.

PromptPanel has been removed as an Inner Class from DragonConsole and added into the dragonconsole.util package as a class of it's own now. PromptPanel will only be used in the future if the Console style is selected that makes use of a two component console style.

TextColor is a new class added to dragonconsole.util. TextColor represents a text style in the console where each "Text Color" is a possible foreground and background. The TextColor object contains a char charCode and a Color color for the TextColor it represents.

CommandProcessor has undergone some minor changes. The method setDragonConsole(DragonConsole) has been changed to install(DragonConsole) and a method has been added so that a DragonConsole can uninstall a CommandProcessor which, in the rare case two CommandProcessors were added to a single console only one would be able to reference the console (the last one added).

git-svn-id: https://dragonconsole.googlecode.com/svn/trunk@15 63ef5802-0434-5e95-cd41-1c6f78d5c2ee