Removed the dc_logo.txt file and replaced it with "logo" and a different text based logo. I also added in a "colors" file that has a basic "for the programmer" explanation of colors. I plan to add in a DemonstrationProcessor that has basic command processing so that programmer can boot up a console by itself and get an idea of what it does and how to perform the tasks.
DragonConsole:
Removed the title global variable, after thinking I realized it really did not make sense for a JPanel to maintain the title since the DragonConsoleFrame was created to handle all basic Frame variables and if implemented into a preexisting interface then a title will already be set. All constructors that dealt with the title have been removed.
Fixed a Logic Error with processing "&&" as "&" or "%%" as "%" (so that if the need to display either of these characters outside of using them to denote colors or input).

FileProcessor:
Added in an (as of yet untested) method for reading a plain text file and returning the contents as a String and added in methods to read files from the JAR as text and returning them. Mostly used for the explanation files for the DemonstrationProcessor and the logo for the consoles default output.

DragonConsoleFrame:
Now has constructors to set a custom title, and no longer pulls the title out of the DragonConsole.

git-svn-id: https://dragonconsole.googlecode.com/svn/trunk@22 63ef5802-0434-5e95-cd41-1c6f78d5c2ee