Changed the default color file to "color_b" for black backgrounds and "color_w" for white, the same with "logo_b" and "logo_w".
Created a "dragonconsole/font" package to contain the font for the Console to ensure the same font on every system (untested).

Debug:
Moved into dragonconsole.util where it made more sense to be.

FileProcessor:
Added in a method to retreive the Font from the .jar

InputController:
Added in a method that works with clearConsole() from DragonConsole to clear all the text in the JTextPane.

DragonConsole:
Added in a method to clear the text, fixed so that if the Console Font is changed it should change for all output text (Not working yet). A fix in PrintDefault to test for a white background, otherwise print the black background.

InputString:
Changed all the "debug.Debug" lines to just Debug and removed the two remaining System.out.println and replaced with Debug.print

DocumentStyler:
Added a method to change the font characteristics for all the styled text in the document already, not working yet.

git-svn-id: https://dragonconsole.googlecode.com/svn/trunk@25 63ef5802-0434-5e95-cd41-1c6f78d5c2ee