This Commit is too large to specify every change (and it's my own fault). The important changes are as follows:
Ownership of the project is now in the hands of 3l33t Software Developers, L.L.C. and no longer in my name however I'm still listed as the @author in the source files. The License tagged to each class has been updated.

Package structure has again changed, this time to reflect the new ownership. All packages are the same, but now part of the com.eleet package. (com.eleet.dragonconsole.*) this is to follow Java package naming standards.

ANSI Color support is functional, however instead of referencing the ANSI class by itself (for color code conversions as of now) you should use the "convertToANSIColors(String)" or "convertToDCColors(String)" of the DragonConsole class to convert DCCC to ANSI or ANSI to DCCC respectively.

I have removed the "colors_w" file and renamed "colors_b" to "colors" as I have made it where it displays readably on both black or white backgrounds. The Logo is still divided into two files so that I can get the Dark over Light effect for the name.

DragonConsole Color Codes (DCCC) now function like ANSI Color Codes do (when adding ANSI support I liked the way the color changes were handled and implemented it for DCCC as well). Meaning, once a color is set, it must be changed, or reset by another code or else output will continue to print with the style. Along with this change came two new Characters for DCCCs which is the "0" (zero) for resetting colors which is used as "&00" (the "&" being the colorCodeChar) and the "-" (hyphen) which is used as a "carry over from previous" character (I.E. If the current Style is "&rb" and you want to change the foreground ("r") to another color without changing the background you would use "&x-" which changes the foreground from "r" to "x" (red to gray) and leaves the background as "b" so the new style is "&xb"). The "0" does not have to be used in doubles, but any 0 in the code will reset the style, ignoring any other characters. Hyphens can be used to carry over the foreground OR the background color, you could technically use "&--" which would do nothing but carry over the current style completely which is automatic and therefore wasteful.

As a final note for this commit, I've been getting weird errors when I do a lot of processing quickly (which confuses me since all the processing is done in the same thread so no Multi-Thread issues should be arising). These errors cycle around as "NullPointerException" with FlowView or "ArrayIndexOutOfBoundsException" with System.arraycopy called from BoxView. I can't seem to diagnose or find any related readings for either and they occur randomly. I can't reliably recreate them on demand for any true debugging of the issue. I should also add the only line of my code that I can trace these errors to (using the printStackTrace()) is the call "this.pack()" in the constructor of DragonConsoleFrame.

git-svn-id: https://dragonconsole.googlecode.com/svn/trunk@36 63ef5802-0434-5e95-cd41-1c6f78d5c2ee