ANSI: Did a bit of code cleaning and added in all the missing comments.
DragonConsoleFrame:
Finally (after forgetting about it since I made the class) added in the setResizable(false); lines since the DragonConsoleFrame is not supposed to be resizable by default. I also added in all the missing comments.

Debug:
Not used at all in this project, but still a possibly valuable utility, now containing comments!

git-svn-id: https://dragonconsole.googlecode.com/svn/trunk@47 63ef5802-0434-5e95-cd41-1c6f78d5c2ee