Added in JOptionPan.showMessageDialog()'s in place of all the black catch blocks so now all errors will at least display that they have failed. These Error Messages are all labeled with a unique number (that should be used when reporting an error so I can find the block that is causing the error). These Error windows were added to the following files: FileProcessor.java DragonConsole.java InputController.java
DragonConsole:
I added Comments to all of the members of this file explaining their purpose. 
I added in @deprecated JavaDOC tags to the three deprecated methods and rewrote them so that they call the proper methods.
Changed all the addTextColor() calls in fillConsoleColors() to refer to the ANSI standard colors instead of creating the same color in two different places.
Added in methods to change/remove/update/clear TextColors added to the console.
Added in a method that gives focus to the inputArea if useInlineInput == false

CommandProcessor:
Added in two methods (convertToANSIColors(String) and convertToDCColors(String)) that reference methods from DragonConsole that may be helpful. (These interfaced methods are in CommandProcessor because I don't want to give the programmer access to the console in CommandProcessor if I don't have to).

DragonConsoleFrame:
Calles the setInputFocus() method of DragonConsole after it calls .pack() in all of it's constructors now.

ANSI:
Renamed all the color variables to meaningful names (i.e. N_0 is now BLACK and B_0 is now INTENSE_BLACK) and made them public so they can be accessed outside of ANSI now.

DocumentStyler:
Removed unused methods (as opposed to deprecating because none of these methods should have ever been referenced outside of DragonConsole) and added a method that will remove a color (and all styles associated with it) from the StyledDocument.

PromptPanel:
PromptPanel got a big overhaul. PromptPanel.PromptLabel is a new class that will render a Prompt with the proper foreground/background colors specified by color codes. I made this changes so that changing the colors of the prompt int he PromptPanel is as ease as changing the colors of text outside of the document. Now, since there are ways to convert from ANSI to DCCCs I saw no need to add ANSI color support here as anything set here that would have ANSI codes can be converted before set.

TextColor:
I commented all the methods and the class in TextColor as well as removed an unnecessary Constructor. I also changed the only public Constructor to throw a TextColor.InvalidCharCodeException if "0" (zero) or "-" (hypen) are passed as a charCode.

git-svn-id: https://dragonconsole.googlecode.com/svn/trunk@51 63ef5802-0434-5e95-cd41-1c6f78d5c2ee