Update to: 3.0.1 (MINOR CHANGE)
FileProcessor:
Removed the final "\n\r" character separating new lines and replaced it simply with "\n" as the JTextPane (in Windows) completely ignores the "\r" but in Mac OS X it makes " " characters in weird places.

InputController:
Change one line, not sure if it was creating a logic error or not but it needed to be consistent with the rest of the class (changed reset() to set rangeStart to -1 instead of 0).

DragonConsole:
Fixed the KeyListener when using inlineInput to prevent protected input from being added to the Previous Command List (which inadvertently made protected input pointless).

DemoProcessor:
Fixed a spelling error.

git-svn-id: https://dragonconsole.googlecode.com/svn/trunk@58 63ef5802-0434-5e95-cd41-1c6f78d5c2ee