Very minor changes:
DragonConsole - Changed a variable name

DemoProcessor - Added in a "help" command with output as well as adding in the Valid argument warning for ANSI which was, for some reason, missing.

InputString - Added in the documentation for a missing @return tag.

Updated the DragonConsole license to reflect the new year, as well as both logos to reflect the new year and version number.

git-svn-id: https://dragonconsole.googlecode.com/svn/trunk@60 63ef5802-0434-5e95-cd41-1c6f78d5c2ee